pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "flacterm"
version = "1.0.0"
description = "Stream FLAC music from the comfort of your Linux terminal"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "GPL-3.0"}
authors = [
    {name = "BRArjun"}
]
maintainers = [
    {name = "BRArjun"}
]
keywords = ["music", "flac", "terminal", "tui", "streaming", "audio", "player"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Players",
    "Topic :: Terminals",
]
dependencies = [
    "textual",
    "rich",
    "requests",
    "python-vlc",
    "lrclibapi",
]

[project.urls]
Homepage = "https://github.com/BRArjun/flacterm"
Documentation = "https://github.com/BRArjun/flacterm#readme"
Repository = "https://github.com/BRArjun/flacterm"
"Bug Tracker" = "https://github.com/BRArjun/flacterm/issues"

[project.scripts]
flacterm = "flacterm.main:main"

[tool.setuptools]
packages = ["flacterm", "flacterm.components", "flacterm.utils"]
include-package-data = true

[tool.setuptools.package-data]
"flacterm" = ["*.txt", "*.md", "*.json"]

[tool.setuptools.exclude-package-data]
"*" = ["*.pyc", "__pycache__/*", "*.pyo", "*.pyd", ".git/*", ".pytest_cache/*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".eggs", "*.egg"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
